+++ easyorange-backend/easyorange-common/src/main/java/com/cartethyia/easyorange/common/util/BizRequire.java
@@ -32,11 +32,23 @@ public class BizRequire {
         }
     }
 
+    public static void notBlank(String str, IResultCode resultCode) {
+        if (str == null || str.isBlank()) {
+            throw BusinessException.of(resultCode);
+        }
+    }
+
     // --- notEmpty ---
 
     public static <T> void notEmpty(Collection<T> collection, String message) {
         if (collection == null || collection.isEmpty()) {
             throw BusinessException.of(message);
+        }
+    }
+
+    public static <T> void notEmpty(Collection<T> collection, IResultCode resultCode) {
+        if (collection == null || collection.isEmpty()) {
+            throw BusinessException.of(resultCode);
         }
     }
 
@@ -46,9 +58,21 @@ public class BizRequire {
         }
     }
 
+    public static void notEmpty(Map<?, ?> map, IResultCode resultCode) {
+        if (map == null || map.isEmpty()) {
+            throw BusinessException.of(resultCode);
+        }
+    }
+
     public static void notEmpty(Object[] array, String message) {
         if (array == null || array.length == 0) {
             throw BusinessException.of(message);
+        }
+    }
+
+    public static void notEmpty(Object[] array, IResultCode resultCode) {
+        if (array == null || array.length == 0) {
+            throw BusinessException.of(resultCode);
         }
     }